Update auto-publishing
Only run auto-publishing on a schedule, so the Telegram messaged doesn't
happen when doing a manual release.
Add a separate workflow file for manual pushes to main. This doesn't
send an automated Telegram message as this will happen when doing a
release.
Pass the Telegram message text to the service.